requirements.txt
Django==5.2.6
scikit-learn==1.7.0
numpy==2.2.2
scipy==1.15.1
pandas==2.2.3
PyMuPDF==1.24.0
python-docx==1.1.0
nltk==3.9.1

